requirements.txt
cos-python-sdk-v5>=1.9.0
requests>=2.25.0 